+++ mobile/kmp/ui/src/commonMain/kotlin/com/atls/hyperion/ui/theme/tokens/elevation/Elevations.kt
@@ -5,30 +5,15 @@ import androidx.compose.ui.unit.Dp
 import com.atls.hyperion.ui.theme.tokens.colors.Colors
 
 class Elevations(colors: Colors) {
-    val xs: PressableElevationStates = Xs(colors.elevation.xs, colors.action.subtle)
-    val sm: PressableElevationStates = Sm(colors.elevation.sm, colors.action.subtle)
-    val md: PressableElevationStates = Md(colors.elevation.md, colors.action.subtle)
-    val lg: PressableElevationStates = Lg(colors.elevation.lg, colors.action.subtle)
+    val xs: PressableElevationStates = Xs(colors.elevation.xs)
+    val sm: PressableElevationStates = Sm(colors.elevation.sm)
+    val md: PressableElevationStates = Md(colors.elevation.md)
+    val lg: PressableElevationStates = Lg(colors.elevation.lg)
     val modal: ElevationStates = Modal(colors.elevation.md)
 
-    private class Xs(color: Color, focusedColor: Color) : PressableElevationStates {
+    private class Xs(color: Color) : PressableElevationStates {
         override val default = elevation(
             drop(color, offsetY = ShadowOffset.xs, blur = ShadowBlur.xs4)
-        )
-        override val focused = elevation(
-            drop(
-                color,
-                offsetY = ShadowOffset.xs,
-                blur = ShadowBlur.xs3,
-                spread = ShadowSpread.xs
-            ),
-            drop(
-                focusedColor,
-                offsetX = ShadowOffset.xs,
-                offsetY = ShadowOffset.md,
-                blur = ShadowBlur.sm,
-                spread = ShadowSpread.negativeSm
-            )
         )
         override val pressed = elevation(
             inner(color, offsetY = ShadowOffset.xs, blur = ShadowBlur.xs5),
@@ -44,24 +29,9 @@ class Elevations(colors: Colors) {
         )
     }
 
-    private class Sm(color: Color, focusedColor: Color) : PressableElevationStates {
+    private class Sm(color: Color) : PressableElevationStates {
         override val default = elevation(
             drop(color, offsetY = ShadowOffset.sm, blur = ShadowBlur.sm)
-        )
-        override val focused = elevation(
-            drop(
-                color,
-                offsetY = ShadowOffset.sm,
-                blur = ShadowBlur.md,
-                spread = ShadowSpread.xs
-            ),
-            drop(
-                focusedColor,
-                offsetX = ShadowOffset.xs,
-                offsetY = ShadowOffset.lg,
-                blur = ShadowBlur.md,
-                spread = ShadowSpread.negativeSm
-            )
         )
         override val pressed = elevation(
             inner(color, offsetY = ShadowOffset.xs, blur = ShadowBlur.xs4),
@@ -77,19 +47,9 @@ class Elevations(colors: Colors) {
         )
     }
 
-    private class Md(color: Color, focusedColor: Color) : PressableElevationStates {
+    private class Md(color: Color) : PressableElevationStates {
         override val default = elevation(
             drop(color, offsetY = ShadowOffset.lg, blur = ShadowBlur.lg)
-        )
-        override val focused = elevation(
-            drop(color, offsetY = ShadowOffset.lg, blur = ShadowBlur.xl),
-            drop(
-                focusedColor,
-                offsetX = ShadowOffset.xs,
-                offsetY = ShadowOffset.xl,
-                blur = ShadowBlur.xl,
-                spread = ShadowSpread.negativeLg
-            )
         )
         override val pressed = elevation(
             inner(color, offsetY = ShadowOffset.xs, blur = ShadowBlur.xs3),
@@ -105,28 +65,13 @@ class Elevations(colors: Colors) {
         )
     }
 
-    private class Lg(color: Color, focusedColor: Color) : PressableElevationStates {
+    private class Lg(color: Color) : PressableElevationStates {
         override val default = elevation(
             drop(
                 color,
                 offsetY = ShadowOffset.xl,
                 blur = ShadowBlur.lg,
                 spread = ShadowSpread.negativeLg
-            )
-        )
-        override val focused = elevation(
-            drop(
-                color,
-                offsetY = ShadowOffset.xl2,
-                blur = ShadowBlur.xl2,
-                spread = ShadowSpread.negativeLg
-            ),
-            drop(
-                focusedColor,
-                offsetX = ShadowOffset.xs,
-                offsetY = ShadowOffset.xl,
-                blur = ShadowBlur.md,
-                spread = ShadowSpread.negativeXs
             )
         )
         override val pressed = elevation(